XD-2462 CI for Kafka Bus
Removed InstanceType.  This allows users to select whatever EC2 instance type they want
Updated to start kafka service on admin
Updated to link kafka to the ZK on the admin
Added checks to verify that kafka is up
Allows user to select what KafkaZK or broker they want to use.